requirements.txt
python-dotenv==0.9.*
requests==2.20.*